TST/REF: method-specific files for indexing methods
